+++ tests/Integration/Request/UpdateUserTest.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Tests\Integration\Request;
 
 use Kreait\Firebase\Auth;
+use Kreait\Firebase\Auth\IdToken;
 use Kreait\Firebase\Request\CreateUser;
 use Kreait\Firebase\Request\UpdateUser;
 use Kreait\Firebase\Tests\IntegrationTestCase;
-use Kreait\Firebase\Util\JSON;
 
 /**
  * @internal
@@ -126,14 +126,13 @@ class UpdateUserTest extends IntegrationTestCase
         $user = $this->auth->updateUser($uid, $request);
         $this->assertEquals($claims, $user->customAttributes);
 
-        // Make sure the custom claims are available in the user's ID token
-        $idTokenResponse = $this->auth->getApiClient()->exchangeCustomTokenForIdAndRefreshToken(
-            $this->auth->createCustomToken($user->uid)
-        );
-        $idToken = $this->auth->verifyIdToken(JSON::decode((string) $idTokenResponse->getBody(), true)['idToken']);
+        $idToken = $this->auth->signInAsUser($user)->idToken();
+        $this->assertInstanceOf(IdToken::class, $idToken);
 
-        $this->assertTrue($idToken->getClaim('admin'));
-        $this->assertSame('1234', $idToken->getClaim('groupId'));
+        $verifiedToken = $this->auth->verifyIdToken($idToken);
+
+        $this->assertTrue($verifiedToken->getClaim('admin'));
+        $this->assertSame('1234', $verifiedToken->getClaim('groupId'));
 
         $this->auth->deleteUser($uid);
     }